backend/requirements.txt
# DarkTrace Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Environment
python-dotenv==1.0.0

# HTTP Client (for future crawler)
httpx==0.25.2

# NLP & ML (basic)
scikit-learn==1.3.2
joblib==1.5.3
numpy==1.26.2
pandas==2.1.3

# Text Processing
spacy==3.7.2

# Graph Analysis (optional)
networkx==3.2.1

# Utilities
python-dateutil==2.8.2

